feat: reorganizar estructura de pagos y añadir tipos de pago
